initially support desktop ui p901 - rename PqServiceHostService into PqServiceHostClient - mv pqTest methods into pqTestService: IPQTestService - mv documentService methods into documentService: IDocumentService - mv evalService methods into evaluationService: IEvaluationService
