+++ src/pqTestConnector/PqServiceHostClient.ts
@@ -5,22 +5,28 @@
  * LICENSE file in the root of this projects source tree.
  */
 
+import * as fs from "fs";
 import * as vscode from "vscode";
 
-import { Credential, ExtensionInfo, IPQTestService } from "../common/PQTestService";
-import { GlobalEventBus, GlobalEvents } from "../GlobalEventBus";
 import {
-    PqServiceHostClientLite,
+    Credential,
+    ExtensionInfo,
+    IPQServiceHostClient,
     PqServiceHostRequestParamBase,
-    PqServiceHostResponseResult,
-} from "./PqServiceHostClientLite";
+} from "../common/PQTestService";
+import { GlobalEventBus, GlobalEvents } from "../GlobalEventBus";
+import { ExtensionConfigurations } from "../constants/PowerQuerySdkConfiguration";
 import { IDisposable } from "../common/Disposable";
 import { PqSdkOutputChannel } from "../features/PqSdkOutputChannel";
+import { PqServiceHostClientLite } from "./PqServiceHostClientLite";
+import { resolveSubstitutedValues } from "../utils/vscodes";
+import { RpcResponseResult } from "./RpcClient";
 import { ValueEventEmitter } from "../common/ValueEventEmitter";
 
 export * from "./PqServiceHostClientLite";
 
-export class PqServiceHostClient extends PqServiceHostClientLite implements IPQTestService, IDisposable {
+export class PqServiceHostClient extends PqServiceHostClientLite implements IPQServiceHostClient, IDisposable {
+    private firstTimeStarted: boolean = true;
     private pingTimer: NodeJS.Timer | undefined = undefined;
 
     public override get pqServiceHostConnected(): boolean {
@@ -70,19 +76,42 @@ export class PqServiceHostClient extends PqServiceHostClientLite implements IPQT
     }
 
     protected override onConnected(): void {
+        super.onConnected();
+
+        // check whether it were the first time staring for the current maybe existing workspace
+        if (this.firstTimeStarted) {
+            // and we also need to ensure we got a valid pq connector mez file
+            const currentPQTestExtensionFileLocation: string | undefined =
+                ExtensionConfigurations.DefaultExtensionLocation;
+
+            const resolvedPQTestExtensionFileLocation: string | undefined = currentPQTestExtensionFileLocation
+                ? resolveSubstitutedValues(currentPQTestExtensionFileLocation)
+                : undefined;
+
+            if (resolvedPQTestExtensionFileLocation && fs.existsSync(resolvedPQTestExtensionFileLocation)) {
+                // trigger one display extension info task to populate modules in the pq-lang ext
+                void this.pqTestService.DisplayExtensionInfo();
+            }
+
+            this.firstTimeStarted = false;
+        }
+
         this.startToSendPingMessages();
     }
 
     protected override onDisconnecting(): void {
+        super.onDisconnecting();
         this.stopSendingPingMessages();
     }
 
     protected override onReconnecting(): void {
+        super.onReconnecting();
+
         // we have already been listening to a service host
         if (this.pingTimer) {
             // there would only one single host expected running per machine
             // thus, we need to shut the existing one down first
-            void this.ForceShutdown();
+            void this.healthService.ForceShutdown();
             // and clear the ping interval handler
             this.stopSendingPingMessages();
         }
@@ -98,7 +127,7 @@ export class PqServiceHostClient extends PqServiceHostClientLite implements IPQT
         } = {},
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
     ): Promise<any> {
-        const responseResultPayload: PqServiceHostResponseResult["Payload"] = await super.requestRemoteRpcMethod(
+        const responseResultPayload: RpcResponseResult["Payload"] = await super.requestRemoteRpcMethod(
             method,
             parameters,
             options,
@@ -117,7 +146,7 @@ export class PqServiceHostClient extends PqServiceHostClientLite implements IPQT
 
     private startToSendPingMessages(): void {
         this.pingTimer = setInterval(() => {
-            void this.Ping();
+            void this.healthService.Ping();
         }, 1950);
     }
 
@@ -132,21 +161,5 @@ export class PqServiceHostClient extends PqServiceHostClientLite implements IPQT
         this.currentExtensionInfos.dispose();
         this.currentCredentials.dispose();
         super.dispose();
-    }
-
-    ForceShutdown(): Promise<number> {
-        return this.requestRemoteRpcMethod("v1/HealthService/Shutdown", [
-            {
-                SessionId: this.sessionId,
-            },
-        ]);
-    }
-
-    Ping(): Promise<number> {
-        return this.requestRemoteRpcMethod("v1/HealthService/Ping", [
-            {
-                SessionId: this.sessionId,
-            },
-        ]);
     }
 }
